+++ packages/core/src/controls/fullscreen-button.ts
@@ -12,7 +12,7 @@ export class FullscreenButtonElement extends PlayerControlElement {
   private handleClick = (): void => {
     if (!this.playerElement) return;
 
-    if (document.fullscreenElement) {
+    if (document.fullscreenElement === this.playerElement) {
       document.exitFullscreen().catch(() => {
         // Ignore fullscreen exit errors
       });
@@ -63,7 +63,9 @@ export class FullscreenButtonElement extends PlayerControlElement {
 
     this.button = document.createElement('button');
     this.button.type = 'button';
+    this.button.setAttribute('part', 'button');
     this.button.setAttribute('aria-label', 'Enter fullscreen');
+    this.button.setAttribute('aria-pressed', 'false');
     this.button.innerHTML = this.getEnterFullscreenIcon();
 
     const shadowRoot = this.attachShadow({ mode: 'open' });
@@ -71,24 +73,26 @@ export class FullscreenButtonElement extends PlayerControlElement {
   }
 
   private getEnterFullscreenIcon(): string {
-    return `<svg class="icon" aria-hidden="true" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+    return `<svg class="icon" part="icon" aria-hidden="true" focusable="false" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
       <path d="M7 14H5v5h5v-2H7v-3zm-2-4h2V7h3V5H5v5zm12 7h-3v2h5v-5h-2v3zM14 5v2h3v3h2V5h-5z"/>
     </svg>`;
   }
 
   private getExitFullscreenIcon(): string {
-    return `<svg class="icon" aria-hidden="true" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+    return `<svg class="icon" part="icon" aria-hidden="true" focusable="false" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
       <path d="M5 16h3v3h2v-5H5v2zm3-8H5v2h5V5H8v3zm6 11h2v-3h3v-2h-5v5zm2-11V5h-2v5h5V8h-3z"/>
     </svg>`;
   }
 
   private updateState(): void {
-    if (document.fullscreenElement) {
+    if (document.fullscreenElement === this.playerElement) {
       this.button.innerHTML = this.getExitFullscreenIcon();
       this.button.setAttribute('aria-label', 'Exit fullscreen');
+      this.button.setAttribute('aria-pressed', 'true');
     } else {
       this.button.innerHTML = this.getEnterFullscreenIcon();
       this.button.setAttribute('aria-label', 'Enter fullscreen');
+      this.button.setAttribute('aria-pressed', 'false');
     }
   }